requirements.txt
openai 
openpyxl 
beautifulsoup4 
python-docx
python-dotenv
requests
fastapi 
uvicorn 
python-multipart
jinja2